+++ src/entities/orders/index.ts
@@ -1,4 +1,4 @@
-import { getCurrentOrder } from '@entities/orders/model/OrderSlice';
+import { getCurrentOrder, setCurrentOrder } from '@entities/orders/model/OrderSlice';
 import OrderCard from './components/orderCard/orderCard';
 import OrderActions from './components/orderCard/orderActions';
-export { OrderCard, OrderActions, getCurrentOrder };
+export { OrderCard, OrderActions, getCurrentOrder, setCurrentOrder };
