--- src/entities/products/components/productList/index.ts
@@ -1,3 +1,0 @@
-import ProductList from './productList';
-
-export default ProductList;
