+++ src/features/editOrderingProductCard/components/editOrderingProductCard.tsx
@@ -1,11 +1,11 @@
-import { transformProductName } from '@entities/products/lib/helpers/transformProductName';
-import { type IOrderProduct } from '@entities/products/model/interfaces/IOrderProduct';
-import { type FC } from 'react';
 import { useForm } from 'react-hook-form';
+import { type FC } from 'react';
+
+import { type IOrderProduct, transformProductName } from '@entities/products';
+import { EditableContent } from '@shared/ui/editableContent';
 import { TextField } from '@shared/ui/textField';
 
 import './editOrderingProductCard.scss';
-import { EditableContent } from '@shared/ui/editableContent/editableContent';
 
 interface EditOrderingProductCardProps {
     product: IOrderProduct
