+++ src/features/popup/components/overlayingPopup/overlayingPopup.scss
