+++ src/features/popup/components/overlayingPopupWithFocusTrap/overlayingPopupWithFocusTrap.scss
@@ -1,0 +1,58 @@
+.popup {
+    &__container {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        position: fixed;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+
+        padding: 36px;
+        box-sizing: border-box;
+
+        &_dropdown {
+            display: block;
+            position: fixed;
+            top: 0;
+            left: 0;
+            right: 0;
+            bottom: 0;
+        }
+    }
+    &__overlay {
+        position: absolute;
+        top: 0;
+        left: 0;
+        right: 0;
+        bottom: 0;
+
+        background-color: rgba(0, 0, 0, 0.6);
+        cursor: pointer;
+    }
+
+    &__message {
+        height: 100%;
+        flex: 1 1 auto;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+    }
+    &__text {
+        text-align: center;
+        font-size: 20px;
+        padding: 0 40px;
+    }
+}
+
+.full-screen-popup {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+
+    padding: 36px;
+    box-sizing: border-box;
+}
