+++ src/features/popup/components/popupCard/popupCard.scss
