--- src/features/popup/overlayingPopup/index.ts
@@ -1,1 +1,0 @@
-export { OverlayingPopupWithFocusTrap } from './overlayingPopupWithFocusTrap';
