+++ src/features/saveOrderToDatabase/index.ts
@@ -1,1 +1,1 @@
-export { SaveOrderToDatabase } from './saveOrderToDatabase';
+export { SaveOrderToDatabase } from './components/saveOrderToDatabase';
